Set up path alias for renderer project
