+++ package.json
@@ -54,7 +54,9 @@
     "cross-env": "catalog:",
     "eslint": "catalog:",
     "eslint-config-prettier": "catalog:",
+    "eslint-import-resolver-typescript": "catalog:",
     "eslint-plugin-better-tailwindcss": "catalog:",
+    "eslint-plugin-import": "catalog:",
     "eslint-plugin-perfectionist": "catalog:",
     "fork-ts-checker-webpack-plugin": "catalog:",
     "glob": "catalog:",
@@ -70,6 +72,7 @@
     "tailwindcss": "catalog:",
     "terser-webpack-plugin": "catalog:",
     "ts-loader": "catalog:",
+    "tsconfig-paths-webpack-plugin": "catalog:",
     "tsdown": "catalog:",
     "typescript": "catalog:",
     "typescript-eslint": "catalog:",
