+++ src/renderer/src/ui/components/button/Button.tsx
@@ -13,9 +13,10 @@ import React, {
   type ReactNode,
 } from "react";
 
-import { joinClasses } from "../../utils/joinClasses";
-import type { XOr } from "../../utils/types";
-import { Icon } from "../icon/Icon";
+import type { XOr } from "@/ui/utils/types";
+
+import { Icon } from "@/ui/components/icon/Icon";
+import { joinClasses } from "@/ui/utils/joinClasses";
 
 export type ButtonType =
   | "primary"
