+++ src/renderer/tsconfig.json
@@ -22,10 +22,10 @@
     "module": "node16", // TODO: change to esnext
     "moduleResolution": "node16", // TODO: change to bundler
     "jsx": "react",
-    "lib": [
-      "ESNext",
-      "DOM",
-      "DOM.Iterable"
-    ]
+    "lib": ["ESNext", "DOM", "DOM.Iterable"],
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["src/*"]
+    }
   }
 }
